requirements.txt
numpy
pyepics
pandas
caproto